+++ src/main/java/org/elasticsearch/gateway/GatewayModule.java
@@ -19,35 +19,21 @@
 
 package org.elasticsearch.gateway;
 
-import com.google.common.collect.ImmutableList;
-import org.elasticsearch.common.inject.AbstractModule;
-import org.elasticsearch.common.inject.Module;
-import org.elasticsearch.common.inject.Modules;
-import org.elasticsearch.common.inject.SpawnModules;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.gateway.local.LocalGatewayModule;
+import org.elasticsearch.common.inject.*;
 
 /**
  *
  */
-public class GatewayModule extends AbstractModule implements SpawnModules {
-
-    public static String GATEWAY_TYPE_SETTING = "gateway.type";
-
-    private final Settings settings;
-
-    public GatewayModule(Settings settings) {
-        this.settings = settings;
-    }
-
-    @Override
-    public Iterable<? extends Module> spawnModules() {
-        Class gateway = settings.getAsClass(GATEWAY_TYPE_SETTING, LocalGatewayModule.class, "org.elasticsearch.gateway.", "GatewayModule");
-        return ImmutableList.of(Modules.createModule(gateway, settings));
-    }
+public class GatewayModule extends AbstractModule {
 
     @Override
     protected void configure() {
         bind(GatewayService.class).asEagerSingleton();
+        bind(Gateway.class).asEagerSingleton();
+        bind(GatewayShardsState.class).asEagerSingleton();
+        bind(TransportNodesListGatewayMetaState.class).asEagerSingleton();
+        bind(GatewayMetaState.class).asEagerSingleton();
+        bind(TransportNodesListGatewayStartedShards.class).asEagerSingleton();
+        bind(LocalAllocateDangledIndices.class).asEagerSingleton();
     }
 }
