+++ src/main/java/org/elasticsearch/gateway/ShardStateInfo.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.gateway.local.state.shards;
+package org.elasticsearch.gateway;
 
 import org.elasticsearch.common.Nullable;
