--- src/main/java/org/elasticsearch/index/gateway/local/LocalIndexGateway.java
@@ -1,58 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.elasticsearch.index.gateway.local;
-
-import org.elasticsearch.common.inject.Inject;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.index.AbstractIndexComponent;
-import org.elasticsearch.index.Index;
-import org.elasticsearch.index.gateway.IndexGateway;
-import org.elasticsearch.index.gateway.IndexShardGateway;
-import org.elasticsearch.index.settings.IndexSettings;
-
-/**
- *
- */
-public class LocalIndexGateway extends AbstractIndexComponent implements IndexGateway {
-
-    @Inject
-    public LocalIndexGateway(Index index, @IndexSettings Settings indexSettings) {
-        super(index, indexSettings);
-    }
-
-    @Override
-    public String type() {
-        return "local";
-    }
-
-    @Override
-    public Class<? extends IndexShardGateway> shardGatewayClass() {
-        return LocalIndexShardGateway.class;
-    }
-
-    @Override
-    public String toString() {
-        return "local";
-    }
-
-    @Override
-    public void close() {
-    }
-}
