+++ src/test/resources/org/elasticsearch/gateway/global-3.st
